Well made steps
Your code seems to be good. I think you did a good job with taking small steps in order to keep a function small and the goal clear. Personally I would of put this code together with de app.js but that is an opinion. If you expand the website this would be a good way to keep this separate.