Painting items
I like the color that you choose for the website. It something and not standard. What I would change is the display of the paintings. I suggest you use a border around the media items, so the overall doensn't look unsteady. I made an example code which gives an idea of what I mean. This is more comforting for the eyes in my opinion. It does need to match with the rest of the styling tho. I think you haven't had enough time for the detail page, but that still needs styling. Overall a good job!